+++ src/Gate.IO.Api/Wallet/Responses/GateWalletSubAccountBalance.cs
@@ -16,4 +16,10 @@ public record GateWalletSubAccountBalance
     /// </summary>
     [JsonProperty("available")]
     public Dictionary<string, decimal> Available { get; set; }
+
+    /// <summary>
+    /// Locked amount by currency
+    /// </summary>
+    [JsonProperty("locking")]
+    public Dictionary<string, decimal> Locked { get; set; }
 }