+++ src/Gate.IO.Api/Wallet/Responses/GateWalletTransferStatus.cs
@@ -9,7 +9,7 @@ public record GateWalletTransferStatus
     /// Order ID
     /// </summary>
     [JsonProperty("tx_id")]
-    public long TransactionId { get; set; }
+    public string TransactionId { get; set; }
 
     /// <summary>
     /// Transfer status, PENDING - in process, SUCCESS - successful transfer, FAIL - failed transfer, PARTIAL_SUCCESS - Partially successful (this status will appear when transferring between sub-subs)